Rename rocfile-private.* to hipfile-private.*
