+++ hipfile/src/amd_detail/hipfile-private.h
