+++ hipfile/test/amd_detail/rocfile.cpp
@@ -15,6 +15,7 @@
 #include "context.h"
 #include "hip.h"
 #include "hipfile.h"
+#include "hipfile-private.h"
 #include "hipfile-warnings.h"
 #include "io.h"
 #include "mbackend.h"
@@ -25,7 +26,6 @@
 #include "mmountinfo.h"
 #include "mstate.h"
 #include "msys.h"
-#include "rocfile-private.h"
 #include "rocfile-test.h"
 #include "state.h"
 #include "sys.h"
